--- example-config.yaml
@@ -1,11 +1,0 @@
-mediaConfing: 
-  host: "localhost"
-  port: 8082
-  prefix: "/media-server"
-  adminPass: "example"
-  tokenLiveTime: 15
-  storageRootPath: "./data/"
-  routes:
-    dataRoute: 
-      name: "/data/"
-      storageRoute: "/"
