+++ internal/domain/fileModel.go
@@ -1,0 +1,6 @@
+package models
+
+type FileModel struct {
+	FileName   string `json:"fileName"`
+	FileBuffer string `json:"fileBuffer"`
+}
