--- src/models/fileModel.go
@@ -1,6 +1,0 @@
-package models
-
-type FileModel struct {
-	FileName   string `json:"fileName"`
-	FileBuffer string `json:"fileBuffer"`
-}
